requirements.txt
boto3
click
pandas
geopandas
plotly
prettytable
pyyaml
pyathena
jupyter
matplotlib
aws-cdk-lib
